test: Build fuzz targets into seperate executables
